Use withKiwtFieldArray HOC
