+++ src/components/LicenseForm/components/index.js
@@ -1,1 +1,2 @@
+export { default as OrganizationsFieldArray } from './OrganizationsFieldArray';
 export { default as TermsListField } from './TermsListField';
